Basic import from OpenDataPhilly to CKAN
This adds a script that performs a basic import from OpenDataPhilly to a
CKAN instance. Not all information is loaded into CKAN since this was
something of a test to see how well the process would work. It seems to
have worked pretty well, so we can probably expand on this for the
future. However, this does result in a CKAN instance with all datasets
from OpenDataPhilly loaded.