Install pdf_view, which moved out of core CKAN
The pdf_preview plugin was formerly built in, but is now a separate
plugin.

Also added some machinery to install simple extensions without having
to make a separate role for each one. "Simple" meaning we just install
it from the repo and add it to the plugins list--no special requirements
or initilization steps needed. Adding separate roles for each of them
feels like too much boilerplate.