Upgrade pip in CKAN virtualenv
CKAN comes with its own virtualenv, which has a version of pip (and wheel
and setuptools) that's too old for some packages. This upgrades it.

It also adds `pip_version` to a `group_vars/all` file and uses the variable
here so that the pip versions installed on the VM and inside the CKAN
virtualenv will always be the same. The VM version might not matter much,
but it's confusing to have a mismatch.