Install showcase extension on CKAN 2.5
In order to migrate the previously built-in Related plugin to the no-longer
built-in Showcase plugin, we must be using CKAN version 2.5

Since we'd prefer to upgrade directly to CKAN 2.8, this is problematic.

The plan to deal with this is to build an AMI using CKAN 2.5 with the
Showcase plugin installed, migrate Related to Showcase, and then use CKAN
2.8 for everything else.

As such, it is not planned for this branch to be merged.